fastapi/requirements.txt
accelerate==1.10.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
etelemetry==0.3.1
fastapi==0.117.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.31
looseversion==1.3.0
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
ml_dtypes==0.5.3
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.2
numpy==2.2.6
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.23.0
openai==1.109.1
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.1
pikepdf==9.11.0
pillow==11.3.0
portalocker==3.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
prov==2.1.1
psutil==7.1.0
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.4
pypandoc==1.15
pyparsing==3.2.5
pypdf==6.1.0
pypdfium2==4.30.0
pyreadline3==3.5.4
pytesseract==0.3.13
pythainlp==5.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==311 ; platform_system == "Windows"
pyxnat==1.6.3
PyYAML==6.0.2
qdrant-client==1.15.1
RapidFuzz==3.14.1
rdflib==7.2.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tika==3.1.0
tiktoken==0.11.0
timm==1.0.20
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traits==7.0.2
transformers==4.56.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.37.0
webencodings==0.5.1
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.20.1
zstandard==0.25.0
